requirements.txt
keybert==0.8.5
openai==1.35.10
pandas==2.2.2
python-dotenv==1.0.1
ruamel.yaml==0.18.6
scikit_learn==1.5.1
